# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VentasFix API",
        "description": "API REST del backoffice de VentasFix para la gestión de usuarios, productos y clientes. Pensada para el consumo desde aplicaciones de terceros (por ejemplo, el sistema de gestión Softland). Todos los endpoints protegidos requieren un token Bearer obtenido en /api/login.",
        "contact": {
            "name": "Felipe Jiménez",
            "email": "felipeandresjb99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar todos los usuarios",
                "operationId": "1517a758326d5c6db5fa5b8b51a020da",
                "responses": {
                    "200": {
                        "description": "Listado de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "19302225f345e9af553e9f0e5c419ecb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut",
                                    "nombre",
                                    "apellido",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "rut": {
                                        "type": "string",
                                        "example": "12.345.678-5"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Felipe"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Jiménez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "felipe@ventasfix.cl"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "clave1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "clave1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario creado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por su ID",
                "operationId": "dcb19f8fda6b09eac753edd96d6af1f1",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario por su ID",
                "operationId": "2243dd2a083b0b3f95947dd1267b28a1",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut",
                                    "nombre",
                                    "apellido",
                                    "email"
                                ],
                                "properties": {
                                    "rut": {
                                        "type": "string",
                                        "example": "12.345.678-5"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Felipe"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Jiménez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "felipe@ventasfix.cl"
                                    },
                                    "password": {
                                        "description": "Opcional. Si se omite, la contraseña no cambia.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "nuevaClave1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nuevaClave1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario actualizado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario por su ID",
                "operationId": "7733142044e974990f1cd75aeb2897f3",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rut": {
                        "type": "string",
                        "example": "12.345.678-5"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Felipe"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Jiménez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "felipe@ventasfix.cl"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduce el token entregado por /api/login como: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}